Add boot / vendor boot image header version 4
The format of boot image v4 is the same as v3.
Vendor boot v4 supports packaging multiple vendor ramdisks by having an
additional section containing a vendor ramdisk table.
The table describes the metadata of each ramdisk, and the offset of each
ramdisk in the vendor ramdisk section.
With these information the bootloader can select a subset of the
ramdisks to load at runtime, and skip loading ramdisks that it knows are
not needed.
The bootloader can also choose to ignore the vendor ramdisk table and
just load the entire vendor ramdisk section. This would have the same
effect as loading all ramdisks in the table.
This lays the foundation for golden boot images.

Bug: 162864255
Test: Presubmit build test
Change-Id: Id3debe7163d802a35aead6f6c076c3ff69c303ee