mkbootimg: Disallow special vendor ramdisk names
Disallow "default" as a vendor ramdisk name, as it has special meaning
for fastboot. For example,

  fastboot flash vendor_boot:default ramdisk.img

Even though the fastboot command only has special meaning when flashing
to a v3 vendor_boot partition, and vendor ramdisk name is a v4 feature,
we still disallow it here to avoid confusion and potential chaos.

Bug: 173654501
Test: Presubmit
Change-Id: I709f19d126c7826806d8523aba98e4356925fc91