mkbootimg: Mark '--gki-*' options as deprecated
The new scheme is to certify the kernel and ramdisk in a
separate pass and generate a piece of certificate, and the certificate
would be placed in the boot_signature slot with the new
'--boot_signature' option.

We still keep the old options working in case we need to certify any old
builds, but mark the options as deprecated so it's clear that we're
moving away from these options.

Bug: 211741246
Test: mkbootimg_test
Change-Id: I69a8e04e80ec200584e0ec5b692c1e8b0fb3193a