Revert "Upgrade zerocopy to 0.7.5"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/940c8417-9273-4d37-a254-9c484747137a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia661eb5348fefb6e2af53d30e7c65cd16798c152