Revert "Upgrade zerocopy to 0.7.5"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fce5bb20-2465-48dc-aa0e-1d254176d4f4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7cc65ae4149b37ef05e9654a0f05655f623f9c3e