Revert "Make mkbootimg a py_binary."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ef17f0f4-fda1-4826-b653-1c01ad9e00a5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id56e30fcfec807b9a45b90436f46d5498d4f92ad
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com