Revert "Make mkbootimg a py_binary."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e5d4e8b5-4830-40d0-ae8f-b0d7f32fe5f1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7b6542cd5f1cb3a708d5daf743a14c5d13f47da1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com